Save current form

+++ public/js/save.js
@@ -33,34 +33,54 @@ function compactJson(json) {
 }
 
 $(document).ready(function() {
-    localStorage.clear();
+    //Clear outstanding local storage.
+    if(localStorage.length != 0){
+        if(confirm('It looks like you have other data locally saved on this browser. Would you like us to clear this data?')){
+            localStorage.clear();
+        }
+    }
+    //Saves form to localstorage on click.
     $("#saveButton").on('click', function() {
-        var name = dataName + counter.toString();
-        saveN($("#mainform").serializeArray(), name);
-        counter++;
+        saveCurrentForm();
     });
+    //Resets the form on click.
     $("#clearForm").on('click', function() {
         $("#mainform")[0].reset();
     });
+    //Pushes the data to the cloud.
     $("#pushData").on('click', function() {
+        //Adds the option to save the current form.
+        if(confirm('Would you like to add the current form to localstorage before saving?')){
+            saveCurrentForm();
+        }
         pushData();
-        counter = 1;
-        localStorage.clear();
     });
+    //Save Current Form
+    function saveCurrentForm(){
+        var name = dataName + counter.toString();
+        saveN($("#mainform").serializeArray(), name);
+        counter++;
+    }
     //Push All Data
     function pushData() { //TODO: Check for connectivity before running.
-        for (var i = 1; i < counter; i++) {
+        var error = false;
+        for (var i = (counter-1); i >= 1; i--) {
             var data = JSON.parse(get('form', i));
             var request = $.ajax({
                 type: 'POST',
                 url: '/pushData',
                 data: data
             });
-            request.done(function(response) {
-                alert(response);
+            request.done(function(response) { //If pushing is successful.
+                // alert(response); //TODO: Add clear message after successful push.
+                localStorage.removeItem('dataName' + i);
+                counter--;
             });
-            request.fail(function(jqXHR, textStatus) {
-                alert(textStatus);
+            request.fail(function(jqXHR, textStatus) { //If pushing is unsuccessful.
+                if(!error){
+                    alert("Pushing was unsuccessful. Please reach out to a qualified individual for assistance. Error Message: " + textStatus);
+                    error = true;
+                }
             });
         }
     }
